Feedback + updates for recursion code
