+++ starky/src/lib.rs
@@ -8,6 +8,7 @@
 pub mod config;
 pub mod constraint_consumer;
 mod get_challenges;
+pub mod permutation;
 pub mod proof;
 pub mod prover;
 pub mod recursive_verifier;
