+++ starky/src/prover.rs
@@ -2,7 +2,6 @@ use std::iter::once;
 
 use anyhow::{ensure, Result};
 use itertools::Itertools;
-use plonky2::field::batch_util::batch_multiply_inplace;
 use plonky2::field::extension_field::Extendable;
 use plonky2::field::field_types::Field;
 use plonky2::field::polynomial::{PolynomialCoeffs, PolynomialValues};
@@ -19,9 +18,9 @@ use rayon::prelude::*;
 
 use crate::config::StarkConfig;
 use crate::constraint_consumer::ConstraintConsumer;
-use crate::get_challenges::get_n_permutation_challenge_sets;
+use crate::permutation::compute_permutation_z_polys;
 use crate::proof::{StarkOpeningSet, StarkProof, StarkProofWithPublicInputs};
-use crate::stark::{PermutationChallenge, PermutationInstance, PermutationPair, Stark};
+use crate::stark::Stark;
 use crate::vars::StarkEvaluationVars;
 
 pub fn prove<F, C, S, const D: usize>(
@@ -61,7 +60,7 @@ where
         "compute trace commitment",
         PolynomialBatch::<F, C, D>::from_values(
             // TODO: Cloning this isn't great; consider having `from_values` accept a reference,
-            // or having `compute_z_poly` read trace values from the `PolynomialBatch`.
+            // or having `compute_permutation_z_polys` read trace values from the `PolynomialBatch`.
             trace_poly_values.clone(),
             rate_bits,
             false,
@@ -76,14 +75,18 @@ where
     challenger.observe_cap(&trace_cap);
 
     // Permutation arguments.
-    let z_commitment = if stark.uses_permutation_args() {
-        let z_polys =
-            compute_z_polys::<F, C, S, D>(&stark, config, &mut challenger, &trace_poly_values);
+    let permutation_zs_commitment = if stark.uses_permutation_args() {
+        let permutation_z_polys = compute_permutation_z_polys::<F, C, S, D>(
+            &stark,
+            config,
+            &mut challenger,
+            &trace_poly_values,
+        );
         timed!(
             timing,
             "compute permutation Z commitments",
             Some(PolynomialBatch::from_values(
-                z_polys,
+                permutation_z_polys,
                 rate_bits,
                 false,
                 config.fri_config.cap_height,
@@ -94,7 +97,7 @@ where
     } else {
         None
     };
-    let permutation_zs_cap = z_commitment
+    let permutation_zs_cap = permutation_zs_commitment
         .as_ref()
         .map(|commit| commit.merkle_tree.cap.clone());
     for cap in &permutation_zs_cap {
@@ -144,11 +147,17 @@ where
         zeta.exp_power_of_2(degree_bits) != F::Extension::ONE,
         "Opening point is in the subgroup."
     );
-    let openings = StarkOpeningSet::new(zeta, g, &trace_commitment, &quotient_commitment);
+    let openings = StarkOpeningSet::new(
+        zeta,
+        g,
+        &trace_commitment,
+        permutation_zs_commitment.as_ref(),
+        &quotient_commitment,
+    );
     challenger.observe_openings(&openings.to_fri_openings());
 
     let initial_merkle_trees = once(&trace_commitment)
-        .chain(z_commitment.as_ref())
+        .chain(permutation_zs_commitment.as_ref())
         .chain(once(&quotient_commitment))
         .collect_vec();
     let fri_params = config.fri_params(degree_bits);
@@ -176,81 +185,6 @@ where
         proof,
         public_inputs: public_inputs.to_vec(),
     })
-}
-
-/// Compute all Z polynomials (for permutation arguments).
-fn compute_z_polys<F, C, S, const D: usize>(
-    stark: &S,
-    config: &StarkConfig,
-    challenger: &mut Challenger<F, C::Hasher>,
-    trace_poly_values: &[PolynomialValues<F>],
-) -> Vec<PolynomialValues<F>>
-where
-    F: RichField + Extendable<D>,
-    C: GenericConfig<D, F = F>,
-    S: Stark<F, D>,
-{
-    let permutation_pairs = stark.permutation_pairs();
-    let permutation_challenge_sets = get_n_permutation_challenge_sets(
-        challenger,
-        config.num_challenges,
-        stark.permutation_batch_size(),
-    );
-
-    let permutation_instances = permutation_pairs
-        .iter()
-        .cartesian_product(0..config.num_challenges)
-        .chunks(stark.permutation_batch_size())
-        .into_iter()
-        .flat_map(|batch| {
-            batch.enumerate().map(|(i, (pair, chal))| {
-                let challenge = permutation_challenge_sets[i].challenges[chal];
-                PermutationInstance { pair, challenge }
-            })
-        })
-        .collect_vec();
-
-    permutation_instances
-        .into_par_iter()
-        .map(|instance| compute_z_poly(instance, trace_poly_values))
-        .collect()
-}
-
-/// Compute a single Z polynomial.
-fn compute_z_poly<F: Field>(
-    instance: PermutationInstance<F>,
-    trace_poly_values: &[PolynomialValues<F>],
-) -> PolynomialValues<F> {
-    let PermutationInstance { pair, challenge } = instance;
-    let PermutationPair {
-        lhs_columns,
-        rhs_columns,
-    } = pair;
-    let PermutationChallenge { beta, gamma } = challenge;
-
-    let degree = trace_poly_values[0].len();
-    let mut reduced_lhs = PolynomialValues::constant(gamma, degree);
-    let mut reduced_rhs = PolynomialValues::constant(gamma, degree);
-
-    let both_cols = lhs_columns.iter().zip_eq(rhs_columns);
-    for ((lhs, rhs), weight) in both_cols.zip(beta.powers()) {
-        reduced_lhs.add_assign_scaled(&trace_poly_values[*lhs], weight);
-        reduced_rhs.add_assign_scaled(&trace_poly_values[*rhs], weight);
-    }
-
-    // Compute the quotients.
-    let reduced_rhs_inverses = F::batch_multiplicative_inverse(&reduced_rhs.values);
-    let mut quotients = reduced_lhs.values;
-    batch_multiply_inplace(&mut quotients, &reduced_rhs_inverses);
-
-    // Compute Z, which contains partial products of the quotients.
-    let mut partial_products = Vec::with_capacity(degree);
-    let mut acc = F::ONE;
-    for q in quotients {
-        partial_products.push(acc);
-        acc *= q;
-    }
-    PolynomialValues::new(partial_products)
 }
 
 /// Computes the quotient polynomials `(sum alpha^i C_i(x)) / Z_H(x)` for `alpha` in `alphas`,
@@ -318,6 +252,7 @@ where
                 public_inputs: &public_inputs,
             };
             stark.eval_packed_base(vars, &mut consumer);
+            // TODO: Add in constraints for permutation arguments.
             // TODO: Fix this once we use a genuine `PackedField`.
             let mut constraints_evals = consumer.accumulators();
             // We divide the constraints evaluations by `Z_H(x)`.
