+++ starky/src/verifier.rs
@@ -89,6 +89,7 @@ where
         l_last,
     );
     stark.eval_ext(vars, &mut consumer);
+    // TODO: Add in constraints for permutation arguments.
     let vanishing_polys_zeta = consumer.accumulators();
 
     // Check each polynomial identity, of the form `vanishing(x) = Z_H(x) quotient(x)`, at zeta.
